requirements.txt
torchvision    
torch    
numpy   
matplotlib      
h5py 
gitpython           
scikit_image
nibabel


